refactor: 타입관련 디렉토리 구조 변경
- @types 대신 types사용
- index.ts내부 타입들 각 도메인 별로 분리
