--- src/@types/form.ts
@@ -1,3 +1,0 @@
-type Validator = (value: string) => void | never;
-
-export type { Validator };
