+++ src/fixtures/data.ts
@@ -1,4 +1,4 @@
-import { Line } from "../@types";
+import { Line } from "../types/line";
 
 const stations = [
   {
