+++ src/modules/auth/thunks.ts
@@ -2,7 +2,7 @@ import { createAsyncThunk, createAction } from "@reduxjs/toolkit";
 
 import { auth } from "../../apis";
 
-import { LoginInfo, SignupInfo } from "../../@types";
+import { LoginInfo, SignupInfo } from "../../types/auth";
 import { ERROR_DURATION } from "../../constants";
 
 const resetError = createAction("[AUTH] RESET_ERROR");
