+++ src/types/auth.ts
@@ -1,0 +1,10 @@
+export interface LoginInfo {
+  email: string;
+  password: string;
+}
+
+export interface SignupInfo {
+  email: string;
+  password: string;
+  age: number;
+}
