+++ src/types/form.ts
@@ -1,0 +1,1 @@
+export type Validator = (value: string) => void | never;
