+++ src/types/route.ts
@@ -1,13 +1,11 @@
-interface RouteShape {
+export interface RouteShape {
   isPrivate: boolean;
   path: string | string[];
   Component: () => JSX.Element;
 }
 
-interface NavLinkShape {
+export interface NavLinkShape {
   isPrivate: boolean;
   to: string;
   title: string;
 }
-
-export type { RouteShape, NavLinkShape };
