+++ src/types/station.ts
@@ -1,0 +1,4 @@
+export interface Station {
+  id: number;
+  name: string;
+}
